student-enrollment-system/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
cx-Oracle==8.3.0
python-dotenv==1.0.0
Werkzeug==2.3.0
SQLAlchemy==2.0.0
